stm32f429/stmpe811: remove platform-specific dependency
Per Jserv's suggestion, this commit removes the platform-specific
dependency from the stmpe-ts driver.

The solution is to invoke the callback function if there is an
exti line on some MCUs.

Enabling the corresponding exti line and clearing the pending exti
interrupt are moved to the stmpe driver instead of the stmpe-ts
driver. It is more reasonable after re-investigating stmpe/stmpe-ts
drivers.